requirements/requirements.txt
django==4.2.16
django-image-cropping==1.7.0
django-haystack==3.2.1
django-extensions==3.2.3
djangorestframework==3.15.2

social-auth-app-django==5.4.2

django-crispy-forms==2.3
crispy-bootstrap4==2024.1

django-braces==1.14.0
django-filter==2.4.0
django-model-utils==4.0.0
drf-spectacular==0.24.2
django-webpack4-loader
python-telegram-bot==20.1

num2words==0.5.13

XlsxWriter==3.2.0

ocrmypdf==14.4.0 # versões acima precisam do ghostscript acima de 9.53.3
pikepdf==8.15.1

dj-database-url==0.5.0
easy-thumbnails==2.8.5

asn1crypto
boto3==1.26.95


#billiard==3.6.3.0

#billiard<4.0,>=3.6.4.0
#celery==5.1.2
#kombu==4.6.8



celery==5.3.6
django-celery-results==2.3.1
git+https://github.com/LeandroJatai/celery-haystack.git
#celery-haystack==0.10

daphne==3.0.2
channels==3.0.5
channels_redis==3.4.1
redis==5.0.1
gunicorn==22.0.0

python-dateutil==2.8.2


jedi==0.17.2
pdfrw
Pillow==10.3.0
psycopg2-binary==2.8.5
pyoai==2.5.0
pypdf4==1.27.0
pysolr==3.9.0
python-decouple==3.1
python-magic==0.4.18
pytz==2021.3
pyyaml==6.0.1
requests==2.32.0
unipath==1.1
WeasyPrint==51
qrcode
reportlab==3.6.13
trml2pdf
kazoo==2.8.0
pymupdf==1.24.7

python-docx==1.1.2

ipython

git+https://github.com/interlegis/django-admin-bootstrapped.git
#git+https://github.com/interlegis/trml2pdf.git
#git+git://github.com/cmjatai/sapl.git@master.para.portal.cmj

#django-extra-views==0.12.0
#django-speedinfo==2.0.1
#redis<=3.3.4
#Whoosh==2.7.4
#django-taggit

#cryptography==3.3.2
